fix host file update on windows
